requirements.txt
# Dinneroo Analysis Pipeline Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
snowflake-connector-python>=3.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# PowerPoint Generation
python-pptx>=0.6.21

# Document Processing
python-docx>=0.8.11

# LLM Integration
google-generativeai>=0.3.0

# Utilities
python-dateutil>=2.8.0
pathlib2>=2.3.0
